assets/gcs_to_database_asset/requirements.txt
google-cloud-storage>=2.0.0
pandas>=1.3.0
sqlalchemy>=2.0.0
